experiments/interviewer-effect/pyproject.toml
[project]
name = "identity-interviewer-effect"
version = "0.1.0"
description = "Tests whether an AI interviewer's theoretical framing causally shifts how subject models describe their own nature"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "polars>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/interviewer_effect"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
